--- doc.go
@@ -1,2 +1,0 @@
-// Package goconsider provides the core of a linter that helps avoiding certain phrases in code.
-package goconsider
