+++ pkg/consider/doc.go
@@ -1,0 +1,2 @@
+// Package consider provides the core of a linter that helps to avoid certain phrases in code.
+package consider
